+++ bin/ocamlformat/main.ml
@@ -84,18 +84,6 @@ let run_action action =
       in
       Result.combine_errors_unit (List.map inputs ~f)
   | Print_config conf -> Conf.print_config conf ; Ok ()
-  | Numeric {kind; file; name= input_name; conf} ->
-      let conf =
-        { conf with
-          opr_opts=
-            { conf.opr_opts with
-              quiet= {v= true; from= conf.opr_opts.quiet.from} } }
-      in
-      let source = source_from_file file in
-      let range = conf.opr_opts.range.v source in
-      Translation_unit.numeric kind ~input_name ~source ~range conf
-      |> List.iter ~f:(fun i -> Stdio.print_endline (Int.to_string i)) ;
-      Ok ()
 ;;
 
 match Bin_conf.action () with